Job context now contains info

+++ src/main/java/com/rbkmoney/scheduledpayoutworker/model/ScheduledJobContext.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class ScheduledJobContext {
 
-    private Long jobId;
+    private Integer jobId;
 
 }
